refactor: Move @Controller annotation methods to ControllerScanner
